Reduce timestamps to uint32 and switched to resolve UUID conflicts using bumps (similar to Solana's PDA)

+++ contracts/IASResolver.sol
@@ -26,7 +26,7 @@ interface IASResolver {
         address recipient,
         bytes calldata schema,
         bytes calldata data,
-        uint256 expirationTime,
+        uint32 expirationTime,
         address msgSender
     ) external payable returns (bool);
 }
