+++ contracts/tests/resolvers/TestASExpirationTimeResolver.sol
@@ -18,7 +18,7 @@ contract TestASExpirationTimeResolver is ASResolver {
         address, /* recipient */
         bytes calldata, /* schema */
         bytes calldata, /* data */
-        uint256 expirationTime,
+        uint32 expirationTime,
         address /* msgSender */
     ) external payable virtual override returns (bool) {
         return expirationTime >= _validAfter;
