+++ contracts/tests/resolvers/TestASPayingResolver.sol
@@ -22,7 +22,7 @@ contract TestASPayingResolver is ASResolver {
         address recipient,
         bytes calldata, /* schema */
         bytes calldata, /* data */
-        uint256, /* expirationTime */
+        uint32, /* expirationTime */
         address /* msgSender */
     ) external payable virtual override returns (bool) {
         payable(recipient).transfer(_incentive);
