+++ contracts/tests/resolvers/TestASTokenResolver.sol
@@ -25,7 +25,7 @@ contract TestASTokenResolver is ASResolver {
         address, /* recipient */
         bytes calldata, /* schema */
         bytes calldata, /* data */
-        uint256, /* expirationTime */
+        uint32, /* expirationTime */
         address msgSender
     ) external payable virtual override returns (bool) {
         _targetToken.safeTransferFrom(msgSender, address(this), _targetAmount);
